Add a generic API to allow session client and session server from hopr-lib
