+++ hopr/hopr-lib/src/lib.rs
@@ -23,22 +23,7 @@ pub mod constants;
 /// Enumerates all errors thrown from this library.
 pub mod errors;
 
-pub use {
-    chain_actions::errors::ChainActionsError,
-    chain_api::config::{
-        Addresses as NetworkContractAddresses, EnvironmentType, Network as ChainNetwork, ProtocolsConfig,
-    },
-    core_transport::{
-        config::{looks_like_domain, HostConfig, HostType},
-        constants::RESERVED_TAG_UPPER_LIMIT,
-        errors::{HoprTransportError, ProtocolError},
-        ApplicationData, HalfKeyChallenge, Health, Keypair, Multiaddr, PathOptions, TicketStatistics, TransportOutput,
-    },
-    hopr_internal_types::prelude::*,
-    hopr_primitive_types::prelude::*,
-    hopr_primitive_types::rlp,
-    hopr_strategy::Strategy,
-};
+mod helpers;
 
 use std::{
     collections::HashMap,
@@ -53,6 +38,8 @@ use futures::{
     Stream, StreamExt,
 };
 use futures_concurrency::stream::StreamExt as _;
+use helpers::WrappedTagBloomFilter;
+use tracing::{debug, error, info, warn};
 
 #[cfg(feature = "runtime-async-std")]
 use async_std::task::{sleep, spawn, JoinHandle};
@@ -76,29 +63,43 @@ use chain_api::{
 use chain_types::chain_events::ChainEventType;
 use chain_types::ContractAddresses;
 use core_path::channel_graph::ChannelGraph;
-use core_transport::{build_network, execute_on_tick, PeerTransportEvent};
-use core_transport::{libp2p::identity::PeerId, Session};
+use core_transport::{execute_on_tick, PeerTransportEvent};
 use core_transport::{ChainKeypair, Hash, HoprTransport, OffchainKeypair};
 use core_transport::{IndexerTransportEvent, Network, PeerEligibility, PeerOrigin};
-use hopr_platform::file::native::{join, read_file, remove_dir_all, write};
+use hopr_crypto_types::prelude::OffchainPublicKey;
+use hopr_db_sql::{
+    accounts::HoprDbAccountOperations,
+    api::{info::SafeInfo, resolver::HoprDbResolverOperations, tickets::HoprDbTicketOperations},
+    channels::HoprDbChannelOperations,
+    db::{HoprDb, HoprDbConfig},
+    info::HoprDbInfoOperations,
+    prelude::{ChainOrPacketKey::ChainKey, DbSqlError, HoprDbPeersOperations},
+    HoprDbAllOperations, HoprDbGeneralModelOperations,
+};
+use hopr_platform::file::native::{join, remove_dir_all};
 use hopr_strategy::strategy::{MultiStrategy, SingularStrategy};
-use tracing::{debug, error, info, warn};
+pub use {
+    chain_actions::errors::ChainActionsError,
+    chain_api::config::{
+        Addresses as NetworkContractAddresses, EnvironmentType, Network as ChainNetwork, ProtocolsConfig,
+    },
+    core_transport::{
+        config::{looks_like_domain, HostConfig, HostType},
+        constants::RESERVED_TAG_UPPER_LIMIT,
+        errors::{HoprTransportError, ProtocolError},
+        libp2p::identity::PeerId,
+        ApplicationData, HalfKeyChallenge, Health, Keypair, Multiaddr, PathOptions, Session as HoprSession,
+        SessionComponents, TicketStatistics, TransportOutput,
+    },
+    hopr_internal_types::prelude::*,
+    hopr_primitive_types::prelude::*,
+    hopr_primitive_types::rlp,
+    hopr_strategy::Strategy,
+};
 
 use crate::constants::{MIN_NATIVE_BALANCE, ONBOARDING_INFORMATION_INTERVAL, SUGGESTED_NATIVE_BALANCE};
 use crate::{config::SafeModule, errors::HoprLibError};
 
-use hopr_db_sql::{
-    accounts::HoprDbAccountOperations,
-    api::{info::SafeInfo, resolver::HoprDbResolverOperations, tickets::HoprDbTicketOperations},
-    db::{HoprDb, HoprDbConfig},
-    info::HoprDbInfoOperations,
-    HoprDbGeneralModelOperations,
-};
-use hopr_db_sql::{channels::HoprDbChannelOperations, HoprDbAllOperations};
-
-use hopr_crypto_types::prelude::OffchainPublicKey;
-use hopr_db_sql::prelude::ChainOrPacketKey::ChainKey;
-use hopr_db_sql::prelude::{DbSqlError, HoprDbPeersOperations};
 #[cfg(all(feature = "prometheus", not(test)))]
 use {
     hopr_metrics::metrics::{MultiGauge, SimpleGauge},
@@ -121,6 +122,15 @@ lazy_static::lazy_static! {
         "Node on-chain and off-chain addresses",
         &["peerid", "address", "safe_address", "module_address"]
     ).unwrap();
+}
+
+/// Interface representing the HOPR server behavior for each incoming session instance
+/// supplied as an argument.
+#[cfg(feature = "session-server")]
+#[async_trait::async_trait]
+pub trait HoprSessionServerActionable {
+    /// Fully process a single HOPR session
+    async fn process(&self, session: HoprSession) -> errors::Result<()>;
 }
 
 /// An enum representing the current state of the HOPR node
@@ -152,8 +162,11 @@ pub struct CloseChannelResult {
 pub enum HoprLibProcesses {
     #[strum(to_string = "libp2p component responsible for the handling of the p2p communication")]
     Swarm,
-    #[strum(to_string = "session unwrapper pairing the session data")]
-    Sessions,
+    #[strum(to_string = "session router pairing the session streams based on the PeerId and ApplicationTag")]
+    SessionsRouter,
+    #[cfg(feature = "session-server")]
+    #[strum(to_string = "session server providing the exit node session stream functionality")]
+    SessionServer,
     #[strum(to_string = "heartbeat component responsible for maintaining the network quality measurements")]
     Heartbeat,
     #[strum(to_string = "tick wake up the strategies to perform an action")]
@@ -358,46 +371,6 @@ where
     .flat_map(futures::stream::iter)
 }
 
-#[derive(Debug, Clone)]
-struct WrappedTagBloomFilter {
-    path: String,
-    tbf: Arc<RwLock<TagBloomFilter>>,
-}
-
-impl WrappedTagBloomFilter {
-    pub fn new(path: String) -> Self {
-        info!("Creating the Bloom filter storage at: {}", path);
-        let tbf = read_file(&path)
-            .and_then(|data| {
-                TagBloomFilter::from_bytes(&data)
-                    .map_err(|e| hopr_platform::error::PlatformError::GeneralError(e.to_string()))
-            })
-            .unwrap_or_else(|_| {
-                debug!("No tag Bloom filter found, using empty");
-                TagBloomFilter::default()
-            });
-
-        Self {
-            path,
-            tbf: Arc::new(RwLock::new(tbf)),
-        }
-    }
-
-    pub fn raw_filter(&self) -> Arc<RwLock<TagBloomFilter>> {
-        self.tbf.clone()
-    }
-
-    pub async fn save(&self) {
-        let bloom = self.tbf.read().await.clone(); // Clone to immediately release the lock
-
-        if let Err(e) = write(&self.path, bloom.to_bytes()) {
-            error!("Tag Bloom filter save failed: {e}")
-        } else {
-            info!("Tag Bloom filter saved successfully")
-        };
-    }
-}
-
 /// Represents the socket behavior of the hopr-lib spawned [`Hopr`] object.
 ///
 /// Provides a read and write stream for Hopr socket recognized data formats.
@@ -458,7 +431,6 @@ pub struct Hopr {
     transport_api: HoprTransport<HoprDb>,
     chain_api: HoprChain<HoprDb>,
     // objects that could be removed pending architectural cleanup ========
-    network: Arc<Network<HoprDb>>,
     db: HoprDb,
     chain_cfg: ChainNetworkConfig,
     channel_graph: Arc<RwLock<core_path::channel_graph::ChannelGraph>>,
@@ -519,17 +491,7 @@ impl Hopr {
         let tbf =
             WrappedTagBloomFilter::new(join(&[&cfg.db.data, "tbf"]).expect("Could not create a tbf storage path"));
 
-        let identity: core_transport::libp2p::identity::Keypair = me.into();
         let my_multiaddresses = vec![multiaddress];
-
-        info!("Creating local network registry and registering own external multiaddresses: {my_multiaddresses:?}",);
-
-        let network = build_network(
-            identity.public().to_peer_id(),
-            my_multiaddresses.clone(),
-            cfg.network_options.clone(),
-            db.clone(),
-        );
 
         let (tx_indexer_events, rx_indexer_events) = async_channel::unbounded::<SignificantChainEvent>();
 
@@ -539,10 +501,10 @@ impl Hopr {
             me,
             me_onchain,
             cfg.transport.clone(),
+            cfg.network_options.clone(),
             cfg.protocol,
             cfg.heartbeat,
             db.clone(),
-            network.clone(),
             channel_graph.clone(),
             my_multiaddresses,
         );
@@ -600,7 +562,6 @@ impl Hopr {
             transport_api: hopr_transport_api,
             chain_api: hopr_chain_api,
             db,
-            network,
             chain_cfg: resolved_environment,
             channel_graph,
             multistrategy: multi_strategy,
@@ -678,7 +639,10 @@ impl Hopr {
         self.cfg.chain.announce
     }
 
-    pub async fn run(&self) -> errors::Result<(HoprSocket, HashMap<HoprLibProcesses, JoinHandle<()>>)> {
+    pub async fn run<#[cfg(feature = "session-server")] T: HoprSessionServerActionable + Clone + Send + 'static>(
+        &self,
+        #[cfg(feature = "session-server")] serve_handler: T,
+    ) -> errors::Result<(HoprSocket, HashMap<HoprLibProcesses, JoinHandle<()>>)> {
         self.error_if_not_in_state(
             HoprState::Uninitialized,
             "Cannot start the hopr node multiple times".into(),
@@ -751,7 +715,7 @@ impl Hopr {
             self.multistrategy.clone(),
             self.channel_graph.clone(),
             self.chain_api.action_state(),
-            self.network.clone(),
+            self.transport_api.network(),
         )
         .await;
 
@@ -813,7 +777,8 @@ impl Hopr {
                 // Self-reference is not needed in the network storage
                 if &peer != self.transport_api.me() {
                     if let Err(e) = self
-                        .network
+                        .transport_api
+                        .network()
                         .add(&peer, PeerOrigin::Initialization, multiaddresses)
                         .await
                     {
@@ -931,15 +896,29 @@ impl Hopr {
             }),
         );
 
+        let (session_tx, _session_rx) = unbounded::<HoprSession>();
+
+        #[cfg(feature = "session-server")]
+        {
+            processes.insert(
+                HoprLibProcesses::SessionServer,
+                spawn(_session_rx.for_each_concurrent(None, move |session| {
+                    let serve_handler = serve_handler.clone();
+                    async move {
+                        let _ = serve_handler.process(session).await;
+                    }
+                })),
+            );
+        }
+
         // TODO: 2.2: add support for transport session handling
-        let (session_tx, _session_rx) = unbounded::<Session>();
         for (id, proc) in self
             .transport_api
             .run(
                 &self.me,
                 &self.me_onchain,
                 String::from(constants::APP_VERSION),
-                self.network.clone(),
+                self.transport_api.network(),
                 self.tbf.raw_filter(),
                 transport_output_tx,
                 on_ack_tkt_tx,
@@ -952,7 +931,7 @@ impl Hopr {
             let nid = match id {
                 core_transport::HoprTransportProcess::Swarm => HoprLibProcesses::Swarm,
                 core_transport::HoprTransportProcess::Heartbeat => HoprLibProcesses::Heartbeat,
-                core_transport::HoprTransportProcess::Sessions => HoprLibProcesses::Sessions,
+                core_transport::HoprTransportProcess::SessionsRouter => HoprLibProcesses::SessionsRouter,
             };
             processes.insert(nid, proc);
         }
@@ -1042,6 +1021,13 @@ impl Hopr {
         self.error_if_not_in_state(HoprState::Running, "Node is not ready for on-chain operations".into())?;
 
         Ok(self.transport_api.ping(peer).await?)
+    }
+
+    #[cfg(feature = "session_client")]
+    pub async fn connect_to(&self, peer: PeerId, path_opts: PathOptions) -> errors::Result<HoprSession> {
+        self.error_if_not_in_state(HoprState::Running, "Node is not ready for on-chain operations".into())?;
+
+        Ok(self.transport_api.new_session(peer, path_opts).await?)
     }
 
     /// Send a message to another peer in the network
