--- transport/api/src/processes/mod.rs
@@ -1,2 +1,0 @@
-/// Background running processes associated with the indexer event processing.
-pub(crate) mod indexer;
